Move things out of 'store' into merlin's connector
