+++ web/app/integration-manager/_merlins-connector/product-details/commands.js
@@ -1,11 +1,10 @@
 import {urlToPathKey} from '../../../utils/utils'
-import {productDetailsParser} from '../../../store/products/parser'
 import {receiveFormInfo} from '../../actions'
 import {receivePdpProductData} from '../../responses'
 
 import {fetchPageData} from '../utils'
 
-import {pdpAddToCartFormParser} from './parsers'
+import {productDetailsParser, pdpAddToCartFormParser} from './parsers'
 
 export const fetchPdpData = (url) => (dispatch) => {
     return dispatch(fetchPageData(url))
