+++ web/app/integration-manager/_merlins-connector/product-details/parsers.js
@@ -1,3 +1,24 @@
+import {extractMagentoJson} from '../../../utils/magento-utils'
+import {getTextFrom} from '../../../utils/parser-utils'
+
+const parseCarouselItems = (magentoObject) => {
+    const carouselSetup = magentoObject
+          .getIn(['[data-gallery-role=gallery-placeholder]', 'mage/gallery/gallery', 'data'])
+          .sortBy((item) => item.get('position'))
+    return carouselSetup.toJS()
+}
+
+export const productDetailsParser = ($, $html) => {
+    const $mainContent = $html.find('.page-main')
+    const magentoObject = extractMagentoJson($html)
+    return {
+        title: getTextFrom($mainContent, '.page-title-wrapper.product .page-title > span'),
+        price: getTextFrom($mainContent, '.product-info-price .price-wrapper .price'),
+        carouselItems: parseCarouselItems(magentoObject),
+        description: getTextFrom($mainContent, '.product.info.detailed .product.attibute.description p')
+    }
+}
+
 export const pdpAddToCartFormParser = ($, $html) => {
     const $mainContent = $html.find('.page-main')
     const $form = $mainContent.find('#product_addtocart_form')
