+++ web/app/store/products/parser.js
@@ -1,13 +1,5 @@
-import {extractMagentoJson} from '../../utils/magento-utils'
 import {getTextFrom, parseTextLink, parseImage} from '../../utils/parser-utils'
 import {urlToPathKey} from '../../utils/utils'
-
-const parseCarouselItems = (magentoObject) => {
-    const carouselSetup = magentoObject
-          .getIn(['[data-gallery-role=gallery-placeholder]', 'mage/gallery/gallery', 'data'])
-          .sortBy((item) => item.get('position'))
-    return carouselSetup.toJS()
-}
 
 export const productListParser = ($, $html) => {
     const $products = $html.find('.item.product-item')
@@ -30,15 +22,4 @@ export const productListParser = ($, $html) => {
         }
     })
     return productMap
-}
-
-export const productDetailsParser = ($, $html) => {
-    const $mainContent = $html.find('.page-main')
-    const magentoObject = extractMagentoJson($html)
-    return {
-        title: getTextFrom($mainContent, '.page-title-wrapper.product .page-title > span'),
-        price: getTextFrom($mainContent, '.product-info-price .price-wrapper .price'),
-        carouselItems: parseCarouselItems(magentoObject),
-        description: getTextFrom($mainContent, '.product.info.detailed .product.attibute.description p')
-    }
 }
